+++ frontend/src/recruiter/components/RecentApplications.jsx
